+++ pwsh/AzGovVizParallel.ps1
@@ -352,10 +352,10 @@ Param
     $Product = 'AzGovViz',
 
     [string]
-    $AzAPICallVersion = '1.1.45',
+    $AzAPICallVersion = '1.1.49',
 
     [string]
-    $ProductVersion = 'v6_major_20221101_1',
+    $ProductVersion = 'v6_major_20221108_1',
 
     [string]
     $GithubRepository = 'aka.ms/AzGovViz',
@@ -504,7 +504,7 @@ Param
     $ShowMemoryUsage,
 
     [int]
-    $CriticalMemoryUsage = 90,
+    $CriticalMemoryUsage = 99,
 
     [switch]
     $DoPSRule,
@@ -544,6 +544,9 @@ Param
 
     [switch]
     $NoNetwork,
+
+    [switch]
+    $ShowRunIdentifier,
 
     #https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/azure-subscription-service-limits#role-based-access-control-limits
     [int]
@@ -7132,11 +7135,10 @@ function processNetwork {
         # }
     }
 
+    $htUnknownTenantsForSubscription = @{}
     foreach ($vnet in $arrayVNets) {
         #peeringsStuff
-
-        #$vnetIdSplit = "/subscriptions/19f26644-2e08-4119-8ade-5e1e93e3dca3/resourceGroups/AzAdvertizer/providers/Microsoft.Network/virtualNetworks/azadvertizer" -split "/"
-        $vnetIdSplit = ($vnet.id -split "/")
+        $vnetIdSplit = ($vnet.id -split '/')
         $subscriptionId = $vnetIdSplit[2]
 
         $subscriptionName = 'n/a'
@@ -7150,16 +7152,50 @@ function processNetwork {
         $vnetResourceGroup = $vnetIdSplit[4]
         if ($vnet.properties.virtualNetworkPeerings.Count -gt 0) {
             foreach ($peering in $vnet.properties.virtualNetworkPeerings) {
-                $remotevnetIdSplit = ($peering.properties.remoteVirtualNetwork.id -split "/")
+                $remotevnetIdSplit = ($peering.properties.remoteVirtualNetwork.id -split '/')
                 $remotesubscriptionId = $remotevnetIdSplit[2]
-
 
                 $remotesubscriptionName = 'n/a'
                 $remoteMGPath = 'n/a'
-                if ($htSubscriptionsMgPath.($subscriptionId)) {
+                $peeringXTenant = 'unknown'
+                if ($htSubscriptionsMgPath.($remotesubscriptionId)) {
+                    $peeringXTenant = 'false'
                     $remotesubHelper = $htSubscriptionsMgPath.($remotesubscriptionId)
                     $remotesubscriptionName = $remotesubHelper.displayName
                     $remoteMGPath = $remotesubHelper.ParentNameChainDelimited
+                }
+                else {
+                    if ($htUnknownTenantsForSubscription.($remotesubscriptionId)) {
+                        $remoteTenantId = $htUnknownTenantsForSubscription.($remotesubscriptionId).TenantId
+                        $remoteMGPath = $remoteTenantId
+                        if ($remoteTenantId -eq $azApiCallConf['checkcontext'].tenant.id) {
+                            $peeringXTenant = 'false'
+                        }
+                        else {
+                            $peeringXTenant = 'true'
+                        }
+                    }
+                    else {
+                        $remotesubscriptionId = '099fca2a-af24-401a-b1ca-42f0ec56e474'
+                        $uri = "$($azAPICallConf['azAPIEndpointUrls'].ARM)/subscriptions/$($remotesubscriptionId)?api-version=2020-01-01"
+                        $remoteTenantId = AzAPICall -AzAPICallConfiguration $azApiCallConf -uri $uri -listenOn 'content' -currentTask "getTenantId for subscriptionId '$($remotesubscriptionId)'"
+                        $arrayRemoteMGPath = @()
+                        foreach ($remoteId in $remoteTenantId) {
+                            $objectGuid = [System.Guid]::empty
+                            if ([System.Guid]::TryParse($remoteId, [System.Management.Automation.PSReference]$ObjectGuid)) {
+                                $arrayRemoteMGPath += $remoteId
+                                if ($remoteId -eq $azApiCallConf['checkcontext'].tenant.id) {
+                                    $peeringXTenant = 'false'
+                                }
+                                else {
+                                    $peeringXTenant = 'true'
+                                }
+                            }
+                            $htUnknownTenantsForSubscription.($remotesubscriptionId) = @{}
+                            $htUnknownTenantsForSubscription.($remotesubscriptionId).TenantId = $arrayRemoteMGPath -join ', '
+                            $remoteMGPath = $arrayRemoteMGPath -join ', '
+                        }
+                    }
                 }
 
                 $remotevnetName = $remotevnetIdSplit[8]
@@ -7189,7 +7225,7 @@ function processNetwork {
                         $remotePeerCompleteVnets = $remotePeering.properties.peerCompleteVnets
                         $remoteRouteServiceVips = $remotePeering.properties.routeServiceVips
                     }
-                    else{
+                    else {
                         $remotePeeringName = 'n/a'
                         $remotePeeringState = 'n/a'
                         $remotePeeringSyncLevel = 'n/a'
@@ -7233,10 +7269,10 @@ function processNetwork {
 
                 $null = $script:arrayVirtualNetworks.Add([PSCustomObject]@{
                         SubscriptionName                                = $subscriptionName
-                        Subscription                                    = ($vnet.id -split "/")[2]
+                        Subscription                                    = ($vnet.id -split '/')[2]
                         MGPath                                          = $MGPath
                         VNet                                            = $vnet.name
-                        VNetId                                            = $vnet.id
+                        VNetId                                          = $vnet.id
                         VNetResourceGroup                               = $vnetResourceGroup
                         Location                                        = $vnet.location
                         AddressSpaceAddressPrefixes                     = $vnet.properties.addressSpace.addressPrefixes
@@ -7249,6 +7285,7 @@ function processNetwork {
                         DdosProtection                                  = $vnet.properties.enableDdosProtection
 
                         PeeringsCount                                   = $vnet.properties.virtualNetworkPeerings.Count
+                        PeeringXTenant                                  = $peeringXTenant
                         PeeringName                                     = $peering.name
                         PeeringState                                    = $peering.properties.peeringState
                         PeeringSyncLevel                                = $peering.properties.peeringSyncLevel
@@ -7297,7 +7334,7 @@ function processNetwork {
         else {
             $null = $script:arrayVirtualNetworks.Add([PSCustomObject]@{
                     SubscriptionName                                = $subscriptionName
-                    Subscription                                    = ($vnet.id -split "/")[2]
+                    Subscription                                    = ($vnet.id -split '/')[2]
                     MGPath                                          = $MGPath
                     VNet                                            = $vnet.name
                     VNetId                                          = $vnet.id
@@ -7314,6 +7351,7 @@ function processNetwork {
                     DdosProtection                                  = $vnet.properties.enableDdosProtection
 
                     PeeringsCount                                   = $vnet.properties.virtualNetworkPeerings.Count
+                    PeeringXTenant                                  = 'n/a'
                     PeeringName                                     = ''
                     PeeringState                                    = ''
                     PeeringSyncLevel                                = ''
@@ -14063,7 +14101,8 @@ extensions: [{ name: 'sort' }]
 
             $htmlSUMMARYALZPolicyVersionChecker = $null
             $exemptionData4CSVExport = [System.Collections.ArrayList]@()
-            $htmlSUMMARYALZPolicyVersionChecker = foreach ($entry in $alzPoliciesInTenant) {
+            $alzPoliciesInTenantSorted = $alzPoliciesInTenant | Sort-Object -Property PolicyName, PolicyId, ALZPolicyName
+            $htmlSUMMARYALZPolicyVersionChecker = foreach ($entry in $alzPoliciesInTenantSorted) {
                 if ([string]::IsNullOrWhiteSpace($entry.AzAdvertizerUrl)) {
                     $link = ''
                 }
@@ -14089,7 +14128,7 @@ extensions: [{ name: 'sort' }]
 
             if (-not $NoCsvExport) {
                 Write-Host "Exporting 'Azure Landing Zones (ALZ) Policy Version Checker' CSV '$($outputPath)$($DirectorySeparatorChar)$($fileName)_ALZPolicyVersionChecker.csv'"
-                $alzPoliciesInTenant | Sort-Object -Property PolicyName, PolicyId | Export-Csv -Path "$($outputPath)$($DirectorySeparatorChar)$($fileName)_ALZPolicyVersionChecker.csv" -Delimiter "$csvDelimiter" -NoTypeInformation
+                $alzPoliciesInTenantSorted | Export-Csv -Path "$($outputPath)$($DirectorySeparatorChar)$($fileName)_ALZPolicyVersionChecker.csv" -Delimiter "$csvDelimiter" -NoTypeInformation
             }
 
             [void]$htmlTenantSummary.AppendLine($htmlSUMMARYALZPolicyVersionChecker)
@@ -20738,10 +20777,12 @@ btn_reset: true, highlight_keywords: true, alternate_rows: true, auto_filter: { 
                 "$($peeringState.Name): $($peeringState.Count)"
             }
 
+            $xTenantPeeringsCount = $vnetPeerings.where({ $_.PeeringXTenant -eq 'true' }).Count
+
             $htmlTableId = 'TenantSummary_VNetPeerings'
             $tfCount = $VNetsPeeringsCount
             [void]$htmlTenantSummary.AppendLine(@"
-<button onclick="loadtf$("func_$htmlTableId")()" type="button" class="collapsible" id="buttonTenantSummary_TenantSummary_VNetPeerings"><i class="padlx fa fa-exchange" aria-hidden="true" style="color: #0078df"></i> <span class="valignMiddle">$VNetsPeeringsCount Virtual Network Peerings - ($($arrayPeeringState -join "$CSVDelimiterOpposite "))</span></button>
+<button onclick="loadtf$("func_$htmlTableId")()" type="button" class="collapsible" id="buttonTenantSummary_TenantSummary_VNetPeerings"><i class="padlx fa fa-exchange" aria-hidden="true" style="color: #0078df"></i> <span class="valignMiddle">$VNetsPeeringsCount Virtual Network Peerings - ($($arrayPeeringState -join "$CSVDelimiterOpposite ")) (Cross Tenant: $($xTenantPeeringsCount))</span></button>
 <div class="content TenantSummary">
 <i class="padlxx fa fa-table" aria-hidden="true"></i> Download CSV <a class="externallink" href="#" onclick="download_table_as_csv_semicolon('$htmlTableId');">semicolon</a> | <a class="externallink" href="#" onclick="download_table_as_csv_comma('$htmlTableId');">comma</a>
 <table id="$htmlTableId" class="summaryTable">
@@ -20762,6 +20803,7 @@ btn_reset: true, highlight_keywords: true, alternate_rows: true, auto_filter: { 
 <th>Connected device</th>
 <th>DDoS</th>
 <th class="uamiresaltbgc">Peerings Count</th>
+<th class="uamiresaltbgc">Peering Cross Tenant</th>
 <th class="uamiresaltbgc">Peering Name</th>
 <th class="uamiresaltbgc">Peering State</th>
 <th class="uamiresaltbgc">Peering Sync Level</th>
@@ -20827,6 +20869,7 @@ btn_reset: true, highlight_keywords: true, alternate_rows: true, auto_filter: { 
                         <td>$($result.ConnectedDevices)</td>
                         <td>$($result.DdosProtection)</td>
                         <td>$($result.PeeringsCount)</td>
+                        <td>$($result.PeeringXTenant)</td>
                         <td>$($result.PeeringName)</td>
                         <td>$($result.PeeringState)</td>
                         <td>$($result.PeeringSyncLevel)</td>
@@ -20909,7 +20952,8 @@ btn_reset: true, highlight_keywords: true, alternate_rows: true, auto_filter: { 
             linked_filters: true,
             col_5: 'select',
             col_13: 'select',
-            col_16: 'select',
+            col_15: 'select',
+
             col_17: 'select',
             col_18: 'select',
             col_19: 'select',
@@ -20918,8 +20962,9 @@ btn_reset: true, highlight_keywords: true, alternate_rows: true, auto_filter: { 
             col_22: 'select',
             col_23: 'select',
             col_24: 'select',
+            col_25: 'select',
 
-            col_27: 'select',
+
             col_28: 'select',
             col_29: 'select',
             col_30: 'select',
@@ -20928,10 +20973,11 @@ btn_reset: true, highlight_keywords: true, alternate_rows: true, auto_filter: { 
             col_33: 'select',
             col_34: 'select',
             col_35: 'select',
+            col_36: 'select',
 
-            col_39: 'select',
-            col_41: 'select',
-            col_51: 'select',
+            col_40: 'select',
+            col_42: 'select',
+            col_52: 'select',
             col_types: [
                 'caseinsensitivestring',
                 'caseinsensitivestring',
@@ -20946,6 +20992,7 @@ btn_reset: true, highlight_keywords: true, alternate_rows: true, auto_filter: { 
                 'number',
                 'number',
                 'number',
+                'caseinsensitivestring',
                 'caseinsensitivestring',
                 'caseinsensitivestring',
                 'caseinsensitivestring',
@@ -26652,7 +26699,7 @@ function stats {
         }
 
         $identifierBase = $hasher512.ComputeHash([System.Text.Encoding]::UTF8.GetBytes($hashUse))
-        $identifier = "$(([System.BitConverter]::ToString($identifierBase)) -replace '-')"
+        $script:statsIdentifier = "$(([System.BitConverter]::ToString($identifierBase)) -replace '-')"
 
         $accountInfo = "$($azAPICallConf['htParameters'].accountType)$($azAPICallConf['htParameters'].userType)"
         if ($azAPICallConf['htParameters'].accountType -eq 'ServicePrincipal' -or $azAPICallConf['htParameters'].accountType -eq 'ManagedService' -or $azAPICallConf['htParameters'].accountType -eq 'ClientAssertion') {
@@ -26690,7 +26737,7 @@ function stats {
             "properties": {
                 "accType": "$($accountInfo)",
                 "azCloud": "$($azAPICallConf['checkContext'].Environment.Name)",
-                "identifier": "$($identifier)",
+                "identifier": "$($statsIdentifier)",
                 "platform": "$($azAPICallConf['htParameters'].CodeRunPlatform)",
                 "productVersion": "$($ProductVersion)",
                 "psAzAccountsVersion": "$($azAPICallConf['htParameters'].AzAccountsVersion)",
@@ -26716,6 +26763,10 @@ function stats {
                 "statsParametersPolicyAtScopeOnly": "$($azAPICallConf['htParameters'].PolicyAtScopeOnly)",
                 "statsParametersRBACAtScopeOnly": "$($azAPICallConf['htParameters'].RBACAtScopeOnly)",
                 "statsParametersDoPSRule": "$($azAPICallConf['htParameters'].DoPSRule)",
+                "statsParametersNoPIMEligibility": "$($NoPIMEligibility)",
+                "statsParametersNoALZPolicyVersionChecker": "$($NoALZPolicyVersionChecker)",
+                "statsParametersNoStorageAccountAccessAnalysis": "$($NoStorageAccountAccessAnalysis)",
+                "statsParametersNoNetwork": "$($NoNetwork)",
                 "statsTry": "$($tryCounter)",
                 "statsDurationProduct": "$($dur)"
             }
@@ -26733,7 +26784,7 @@ function stats {
     }
     else {
         #noStats
-        $identifier = (New-Guid).Guid
+        $script:statsIdentifier = (New-Guid).Guid
         $tryCounter = 0
         do {
             if ($tryCounter -gt 0) {
@@ -26753,7 +26804,7 @@ function stats {
             "name": "$($Product)",
             "ver": 2,
             "properties": {
-                "identifier": "$($identifier)",
+                "identifier": "$($statsIdentifier)",
                 "statsTry": "$($tryCounter)"
             }
         }
@@ -31883,7 +31934,8 @@ if ($azAPICallConf['htParameters'].HierarchyMapOnly -eq $false) {
     $policyBaseQuerySubscriptions = $policyBaseQuery.where({ -not [String]::IsNullOrEmpty($_.SubscriptionId) } )
     $policyBaseQueryManagementGroups = $policyBaseQuery.where({ [String]::IsNullOrEmpty($_.SubscriptionId) } )
     $policyPolicyBaseQueryScopeInsights = ($policyBaseQuery | Select-Object Mg*, Subscription*, PolicyAssignmentAtScopeCount, PolicySetAssignmentAtScopeCount, PolicyAndPolicySetAssignmentAtScopeCount, PolicyAssignmentLimit -Unique)
-    $policyBaseQueryUniqueAssignments = $policyBaseQuery | Select-Object -Property Policy* | Sort-Object -Property PolicyAssignmentId -Unique
+    #$policyBaseQueryUniqueAssignments = $policyBaseQuery | Select-Object -Property Policy* | Sort-Object -Property PolicyAssignmentId -Unique
+    $policyBaseQueryUniqueAssignments = $policyBaseQuery | Sort-Object -Property PolicyAssignmentId -Unique | Select-Object -Property Policy*
     $policyAssignmentsOrphaned = $policyBaseQuery.where({ $_.PolicyAvailability -eq 'na' } ) | Sort-Object -Property PolicyAssignmentId -Unique
     $policyAssignmentsOrphanedCount = $policyAssignmentsOrphaned.Count
     Write-Host "  $policyAssignmentsOrphanedCount orphaned Policy assignments found"
@@ -31904,7 +31956,8 @@ if ($azAPICallConf['htParameters'].HierarchyMapOnly -eq $false) {
     }
 
     $policyPolicySetBaseQueryUniqueAssignments = $policyBaseQueryUniqueAssignments.where({ $_.PolicyVariant -eq 'PolicySet' } )
-    $policyBaseQueryUniqueCustomDefinitions = ($policyBaseQuery.where({ $_.PolicyType -eq 'Custom' } )) | Select-Object PolicyVariant, PolicyDefinitionId -Unique
+    #$policyBaseQueryUniqueCustomDefinitions = ($policyBaseQuery.where({ $_.PolicyType -eq 'Custom' } )) | Select-Object PolicyVariant, PolicyDefinitionId -Unique
+    $policyBaseQueryUniqueCustomDefinitions = ($policyBaseQuery.where({ $_.PolicyType -eq 'Custom' } )) | Sort-Object -Property PolicyVariant, PolicyDefinitionId -Unique | Select-Object PolicyVariant, PolicyDefinitionId
     $policyPolicyBaseQueryUniqueCustomDefinitions = ($policyBaseQueryUniqueCustomDefinitions.where({ $_.PolicyVariant -eq 'Policy' } )).PolicyDefinitionId
     $policyPolicySetBaseQueryUniqueCustomDefinitions = ($policyBaseQueryUniqueCustomDefinitions.where({ $_.PolicyVariant -eq 'PolicySet' } )).PolicyDefinitionId
 
@@ -31920,7 +31973,8 @@ if ($azAPICallConf['htParameters'].HierarchyMapOnly -eq $false) {
     }
 
     $blueprintBaseQuery = ($newTable | Select-Object mgid, SubscriptionId, Blueprint*).where({ -not [String]::IsNullOrEmpty($_.BlueprintName) } )
-    $mgsAndSubs = (($optimizedTableForPathQuery.where({ $_.mgId -ne '' -and $_.Level -ne '0' } )) | Select-Object MgId, SubscriptionId -Unique)
+    #$mgsAndSubs = (($optimizedTableForPathQuery.where({ $_.mgId -ne '' -and $_.Level -ne '0' } )) | Select-Object MgId, SubscriptionId -Unique)
+    $mgsAndSubs = (($optimizedTableForPathQuery.where({ $_.mgId -ne '' -and $_.Level -ne '0' } )) | Sort-Object -Property MgId, SubscriptionId -Unique | Select-Object MgId, SubscriptionId)
 
     #region create array Policy definitions
     $tenantAllPoliciesCount = (($htCacheDefinitionsPolicy).Values).count
@@ -32096,7 +32150,8 @@ if ($azAPICallConf['htParameters'].HierarchyMapOnly -eq $false) {
     #region summarizeDataCollectionResults
     $startSummarizeDataCollectionResults = Get-Date
     Write-Host 'Summary data collection'
-    $mgsDetails = ($optimizedTableForPathQueryMg | Select-Object Level, MgId -Unique)
+    #$mgsDetails = ($optimizedTableForPathQueryMg | Select-Object Level, MgId -Unique)
+    $mgsDetails = ($optimizedTableForPathQueryMg | Sort-Object -Property Level, MgId -Unique | Select-Object Level, MgId)
     $mgDepth = ($mgsDetails.Level | Measure-Object -Maximum).Maximum
     $totalMgCount = ($mgsDetails).count
     $totalSubCount = ($optimizedTableForPathQuerySub).count
@@ -32984,3 +33039,10 @@ if ($azGovVizNewerVersionAvailable) {
     }
 }
 #endregion infoNewAzGovVizVersionAvailable
+
+#region runIdentifier
+if ($ShowRunIdentifier) {
+    Write-Host "AzGovViz run identifier: '$($statsIdentifier)'"
+}
+#endregion runIdentifier
+